move SelectionPopup to teams/common
